nn/proj/pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "tensorflow-serving-api>=2.19.1",
    "tf-keras>=2.20.1",
    "uvicorn>=0.37.0",
]
